Refactor: 메시지 가져오는 코드 기능별 모듈화
